ISA_synthesized/requirements.txt
# Core Application
python-dotenv>=1.0.1,<2
fastapi>=0.116.1,<0.117
uvicorn[standard]>=0.35.0,<0.36
requests>=2.32.4,<3
beautifulsoup4>=4.13.4,<5
neo4j>=5.28.0,<6
supabase>=2.18.1,<3
schedule>=1.2.1,<2
sse-starlette>=2.1.2,<3
anyio>=4.4.0
dspy-ai>=3.0.1
langgraph>=0.0.50
PyJWT>=2.9.0,<3

# ETL Pipeline Dependencies
pandas
pyarrow
tqdm
lxml
rdflib
jsonschema
pydantic
langdetect
regex
unstructured
pdfminer.six
PyYAML